# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kuben API",
    "description": "Kuben — Kubernetes-native PaaS control plane.",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.1.2"
  },
  "paths": {
...
    "/api/v1/projects/{project}/environments/{environment}/templates/{template}": {
      "post": {
        "tags": [
          "templates"
        ],
        "summary": "Deploy a template into an environment.",
        "operationId": "deployTemplate",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment short name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "template",
            "in": "path",
            "description": "Template id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeployTemplate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeployedTemplate"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppDto": {
        "type": "object",
        "required": [
          "name",
          "project",
          "environment",
          "namespace",
          "ready",
          "processes",
          "env",
          "domains",
          "volumes"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "project": {
            "type": "string"
          },
          "environment": {
            "type": "string",
            "description": "Environment short name."
          },
          "namespace": {
            "type": "string"
          },
          "image": {
            "type": [
              "string",
              "null"
            ]
          },
          "git_repo": {
            "type": [
              "string",
              "null"
            ]
          },
          "url": {
            "type": [
              "string",
              "null"
            ]
          },
          "ready": {
            "type": "boolean"
          },
          "reason": {
            "type": [
              "string",
              "null"
            ],
            "description": "`Available`, `Progressing`, `RolloutFailed`, `Scheduled`, `AwaitingBuild`, ..."
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "processes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProcessDto"
            }
          },
          "env": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnvVarDto"
            },
            "description": "Values are only included in the app detail, for callers allowed to read secrets."
          },
          "domains": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "volumes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VolumeDto"
            }
          },
          "created_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "exposure": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ExposureDto",
                "description": "Whether the app can be reached through the gateway, apart from\nwhether it runs (null while it has no route)."
              }
            ]
          },
          "paused": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why delivery is paused, while it is: new runs wait."
          }
        }
      },
...
      "DeployTemplate": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "App name (also the in-cluster hostname of TCP services).",
            "example": "db"
          }
        }
      },
      "DeployedTemplate": {
        "type": "object",
        "required": [
          "app",
          "credentials_secret",
          "connection_keys"
        ],
        "properties": {
          "app": {
            "$ref": "#/components/schemas/AppDto"
          },
          "credentials_secret": {
            "type": "string",
            "description": "Secret with the generated credentials, e.g. `db-credentials`."
          },
          "connection_keys": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Reference it from other apps as `KEY=@<secret>/<key>`."
          }
        }
      },
...
      "EnvVarDto": {
        "type": "object",
        "description": "Environment variable: a plain `value` or a `secret` reference.",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "LOG_LEVEL"
          },
          "value": {
            "type": [
              "string",
              "null"
            ]
          },
          "secret": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SecretRef"
              }
            ]
          }
        }
      },
...
      "ExposureDto": {
        "type": "object",
        "description": "How an app is reached through the gateway.",
        "required": [
          "hosts"
        ],
        "properties": {
          "routed": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "The gateway accepted the route and resolved its references; null\nuntil a gateway controller answered."
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "hosts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HostDto"
            }
          }
        }
      },
...
      "HostDto": {
        "type": "object",
        "description": "One hostname of an app.",
        "required": [
          "host",
          "tls"
        ],
        "properties": {
          "host": {
            "type": "string"
          },
          "tls": {
            "type": "string",
            "description": "`auto` (certificate from the cluster issuer), `secret` (the app's own\ncertificate) or `none` (plain HTTP)."
          },
          "certificate_ready": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "For `auto` hosts with a certificate of their own: whether it is issued."
          },
          "certificate_message": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "Problem Details body.",
        "required": [
          "code",
          "title",
          "status"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, e.g. `forbidden`.",
            "example": "forbidden"
          },
          "title": {
            "type": "string",
            "description": "Short human readable title."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status.",
            "minimum": 0
          },
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "description": "Details, if safe to expose."
          }
        }
      },
      "ProcessDto": {
        "type": "object",
        "required": [
          "name",
          "command",
          "size",
          "min_replicas",
          "max_replicas",
          "protocol"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "command": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "port": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "string"
          },
          "min_replicas": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "max_replicas": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "schedule": {
            "type": [
              "string",
              "null"
            ],
            "description": "Cron expression of scheduled processes."
          },
          "protocol": {
            "type": "string",
            "description": "`http` or `tcp`."
          }
        }
      },
...
      "SecretRef": {
        "type": "object",
        "required": [
          "name",
          "key"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "key": {
            "type": "string"
          }
        }
      },
...
      "VolumeDto": {
        "type": "object",
        "description": "Persistent volume (kept when the app is deleted, unless requested).",
        "required": [
          "name",
          "mount_path"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "data"
          },
          "mount_path": {
            "type": "string",
            "example": "/data"
          },
          "size": {
            "type": "string",
            "description": "Capacity such as `5Gi`. Can grow, never shrink.",
            "example": "5Gi"
          }
        }
      },
...
    }
  },
...
}